+++ src/components/app-hero/composables/useHeroRouteDependentSettings.ts
